Support local scoping for functions (with bugs)
Can now do:
g x = f 1 + x

But not (yet):
g x = f x + 1

To support the later need some way of updating the precedence of a node.
f and x are both references, so f is evaluated first. But since f
happens to be a function taking x, we need to evaluate x first. But
we only know f is a function when we're already evaluating it.